Added the SimpleHammer class
For this commit I did the following added the SimpleHammer class,
modified WeaponFactory to give a simple hammer with the string
"simpleHammer", and modified RoleGame class to test simple hammer.
In the RoleGame class I added a test function that prints the amount of
damage a weapon does and how much damage it does dased off the armor
value given. I test the hammer for 29 armor and 30 armor.